+++ services/src/main/java/org/keycloak/broker/oidc/OAuth2IdentityProviderConfig.java
@@ -18,16 +18,22 @@ package org.keycloak.broker.oidc;
 
 import static org.keycloak.common.util.UriUtils.checkUrl;
 
+import org.keycloak.OAuth2Constants;
 import org.keycloak.common.enums.SslRequired;
 import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
 import org.keycloak.protocol.oidc.OIDCLoginProtocol;
 
+import java.util.Arrays;
+
 /**
  * @author Pedro Igor
  */
 public class OAuth2IdentityProviderConfig extends IdentityProviderModel {
+
+    public static final String PKCE_ENABLED = "pkceEnabled";
+    public static final String PKCE_METHOD = "pkceMethod";
 
     public OAuth2IdentityProviderConfig(IdentityProviderModel model) {
         super(model);
@@ -125,6 +131,22 @@ public class OAuth2IdentityProviderConfig extends IdentityProviderModel {
        getConfig().put("forwardParameters", forwardParameters);
     }
 
+    public boolean isPkceEnabled() {
+        return Boolean.parseBoolean(getConfig().getOrDefault(PKCE_ENABLED, "false"));
+    }
+
+    public void setPkceEnabled(boolean enabled) {
+        getConfig().put(PKCE_ENABLED, String.valueOf(enabled));
+    }
+
+    public String getPkceMethod() {
+        return getConfig().get(PKCE_METHOD);
+    }
+
+    public String setPkceMethod(String method) {
+        return getConfig().put(PKCE_METHOD, method);
+    }
+
     @Override
     public void validate(RealmModel realm) {
         SslRequired sslRequired = realm.getSslRequired();
@@ -132,5 +154,13 @@ public class OAuth2IdentityProviderConfig extends IdentityProviderModel {
         checkUrl(sslRequired, getAuthorizationUrl(), "authorization_url");
         checkUrl(sslRequired, getTokenUrl(), "token_url");
         checkUrl(sslRequired, getUserInfoUrl(), "userinfo_url");
+
+
+        if (isPkceEnabled()) {
+            String pkceMethod = getPkceMethod();
+            if (!Arrays.asList(OAuth2Constants.PKCE_METHOD_PLAIN, OAuth2Constants.PKCE_METHOD_S256).contains(pkceMethod)) {
+                throw new IllegalArgumentException("PKCE Method not supported: " + pkceMethod);
+            }
+        }
     }
 }
