+++ services/src/main/java/org/keycloak/broker/oidc/OIDCIdentityProviderConfig.java
@@ -34,8 +34,6 @@ public class OIDCIdentityProviderConfig extends OAuth2IdentityProviderConfig {
 
     public static final String USE_JWKS_URL = "useJwksUrl";
     public static final String VALIDATE_SIGNATURE = "validateSignature";
-    public static final String PKCE_ENABLED = "pkceEnabled";
-    public static final String PKCE_METHOD = "pkceMethod";
 
     public OIDCIdentityProviderConfig(IdentityProviderModel identityProviderModel) {
         super(identityProviderModel);
@@ -135,34 +133,11 @@ public class OIDCIdentityProviderConfig extends OAuth2IdentityProviderConfig {
         }
     }
 
-    public boolean isPkceEnabled() {
-        return Boolean.parseBoolean(getConfig().getOrDefault(PKCE_ENABLED, "false"));
-    }
-
-    public void setPkceEnabled(boolean enabled) {
-        getConfig().put(PKCE_ENABLED, String.valueOf(enabled));
-    }
-
-    public String getPkceMethod() {
-        return getConfig().get(PKCE_METHOD);
-    }
-
-    public String setPkceMethod(String method) {
-        return getConfig().put(PKCE_METHOD, method);
-    }
-
-    @Override 
+    @Override
     public void validate(RealmModel realm) {
         super.validate(realm);
         SslRequired sslRequired = realm.getSslRequired();
         checkUrl(sslRequired, getJwksUrl(), "jwks_url");
         checkUrl(sslRequired, getLogoutUrl(), "logout_url");
-
-        if (isPkceEnabled()) {
-            String pkceMethod = getPkceMethod();
-            if (!Arrays.asList(OAuth2Constants.PKCE_METHOD_PLAIN, OAuth2Constants.PKCE_METHOD_S256).contains(pkceMethod)) {
-                throw new IllegalArgumentException("PKCE Method not supported: " + pkceMethod);
-            }
-        }
     }
 }
